entity_relation -> entity_property
