+++ adapters/bevy/shared/src/lib.rs
@@ -1,8 +1,8 @@
 pub use naia_shared::{
     sequence_greater_than, BitReader, BitWrite, BitWriter, Channel, ChannelDirection, ChannelKind,
     ChannelMode, ComponentKind, ComponentKinds, ComponentUpdate, ConstBitLength, DiffMask,
-    EntityAndGlobalEntityConverter, EntityDoesNotExistError, EntityProperty, EntityRelation,
-    GlobalEntity, LinkConditionerConfig, LocalEntity, LocalEntityAndGlobalEntityConverter,
+    EntityAndGlobalEntityConverter, EntityDoesNotExistError, EntityProperty, GlobalEntity,
+    LinkConditionerConfig, LocalEntity, LocalEntityAndGlobalEntityConverter,
     LocalEntityAndGlobalEntityConverterMut, MessageBevy as Message, MessageBuilder,
     MessageContainer, MessageKind, MessageKinds, Named, OwnedBitReader, Property, PropertyMutate,
     PropertyMutator, Random, ReliableSettings, ReplicaDynMut, ReplicaDynRef,
