+++ shared/src/world/component/entity_property.rs
@@ -1,40 +1,34 @@
+use log::warn;
 use std::hash::Hash;
 
-use naia_serde::{BitCounter, BitReader, BitWrite, BitWriter, Serde, SerdeErr};
+use naia_serde::{BitCounter, BitReader, BitWrite, Serde, SerdeErr};
 
-use crate::world::{
-    component::{property::Property, property_mutate::PropertyMutator},
-    entity::{
-        entity_converters::{EntityAndGlobalEntityConverter, LocalEntityAndGlobalEntityConverter},
-        global_entity::GlobalEntity,
-        local_entity::LocalEntity,
+use crate::world::entity::{
+    entity_converters::{
+        EntityAndGlobalEntityConverter, LocalEntityAndGlobalEntityConverter,
+        LocalEntityAndGlobalEntityConverterMut,
     },
+    global_entity::GlobalEntity,
+    local_entity::LocalEntity,
 };
 
 #[derive(Clone)]
+enum EntityRelation {
+    HostOwned(HostOwnedRelation),
+    RemoteOwned(RemoteOwnedRelation),
+    RemoteWaiting(RemoteWaitingRelation),
+}
+
+#[derive(Clone)]
 pub struct EntityProperty {
-    global_entity_prop: Property<Option<GlobalEntity>>,
+    inner: EntityRelation,
 }
 
 impl EntityProperty {
-    pub fn new(mutator_index: u8) -> Self {
+    pub fn new() -> Self {
         Self {
-            global_entity_prop: Property::<Option<GlobalEntity>>::new(None, mutator_index),
+            inner: EntityRelation::HostOwned(HostOwnedRelation::new()),
         }
-    }
-
-    pub fn new_empty() -> Self {
-        Self {
-            global_entity_prop: Property::<Option<GlobalEntity>>::new(None, 0),
-        }
-    }
-
-    pub fn mirror(&mut self, other: &EntityProperty) {
-        *self.global_entity_prop = other.global_entity();
-    }
-
-    pub fn global_entity(&self) -> Option<GlobalEntity> {
-        *self.global_entity_prop
     }
 
     // Serialization / deserialization
@@ -42,106 +36,177 @@ impl EntityProperty {
     pub fn write(
         &self,
         writer: &mut dyn BitWrite,
-        converter: &dyn LocalEntityAndGlobalEntityConverter,
+        converter: &mut dyn LocalEntityAndGlobalEntityConverterMut,
     ) {
-        if let Some(global_entity) = *self.global_entity_prop {
-            if let Ok(local_entity) = converter.global_entity_to_local_entity(&global_entity) {
-                // Must reverse the LocalEntity because the Host<->Remote
-                // relationship inverts after this data goes over the wire
-                let reversed_local_entity = local_entity.to_reversed();
-
-                true.ser(writer);
-                reversed_local_entity.owned_ser(writer);
-                return;
+        match &self.inner {
+            EntityRelation::HostOwned(inner) => {
+                inner.write(writer, converter);
+            }
+            EntityRelation::RemoteOwned(_) | EntityRelation::RemoteWaiting(_) => {
+                panic!("Remote Relations should never be written.");
             }
         }
-        false.ser(writer);
     }
 
-    pub fn bit_length(&self, converter: &dyn LocalEntityAndGlobalEntityConverter) -> u32 {
-        let mut bit_counter = BitCounter::new(0, 0, u32::MAX);
-        self.write(&mut bit_counter, converter);
-        return bit_counter.bits_needed();
+    pub fn bit_length(&self, converter: &mut dyn LocalEntityAndGlobalEntityConverterMut) -> u32 {
+        match &self.inner {
+            EntityRelation::HostOwned(inner) => inner.bit_length(converter),
+            EntityRelation::RemoteOwned(_) | EntityRelation::RemoteWaiting(_) => {
+                panic!(
+                    "Remote Relations should never be written, so no need for their bit length."
+                );
+            }
+        }
     }
 
-    pub fn new_read(
+    pub fn read(
         reader: &mut BitReader,
-        mutator_index: u8,
         converter: &dyn LocalEntityAndGlobalEntityConverter,
     ) -> Result<Self, SerdeErr> {
         let exists = bool::de(reader)?;
         if exists {
             let local_entity = LocalEntity::owned_de(reader)?;
             if let Ok(global_entity) = converter.local_entity_to_global_entity(&local_entity) {
-                let mut new_prop = Self::new(mutator_index);
-                *new_prop.global_entity_prop = Some(global_entity);
-                Ok(new_prop)
+                let mut new_impl = RemoteOwnedRelation::new();
+                new_impl.global_entity = Some(global_entity);
+
+                let new_self = Self {
+                    inner: EntityRelation::RemoteOwned(new_impl),
+                };
+
+                Ok(new_self)
             } else {
-                panic!(
-                    "Could not find GlobalEntity to associate with incoming EntityProperty value!"
-                );
+                let new_impl = RemoteWaitingRelation::new(local_entity);
+
+                let new_self = Self {
+                    inner: EntityRelation::RemoteWaiting(new_impl),
+                };
+
+                Ok(new_self)
             }
         } else {
-            let mut new_prop = Self::new(mutator_index);
-            *new_prop.global_entity_prop = None;
-            Ok(new_prop)
-        }
-    }
+            let mut new_impl = RemoteOwnedRelation::new();
+            new_impl.global_entity = None;
 
-    pub fn read_write(reader: &mut BitReader, writer: &mut BitWriter) -> Result<(), SerdeErr> {
-        let exists = bool::de(reader)?;
-        exists.ser(writer);
-        if exists {
-            let local_entity = LocalEntity::owned_de(reader)?;
-            local_entity.owned_ser(writer);
-        }
-        Ok(())
-    }
+            let new_self = Self {
+                inner: EntityRelation::RemoteOwned(new_impl),
+            };
 
-    pub fn read(
-        &mut self,
-        reader: &mut BitReader,
-        converter: &dyn LocalEntityAndGlobalEntityConverter,
-    ) -> Result<(), SerdeErr> {
-        let exists = bool::de(reader)?;
-        if exists {
-            let local_entity = LocalEntity::owned_de(reader)?;
-            if let Ok(global_entity) = converter.local_entity_to_global_entity(&local_entity) {
-                *self.global_entity_prop = Some(global_entity);
-            } else {
-                panic!("Could not find Entity to associate with incoming EntityProperty value!");
-            }
-        } else {
-            *self.global_entity_prop = None;
+            Ok(new_self)
         }
-        Ok(())
-    }
-
-    // Comparison
-
-    pub fn equals(&self, other: &EntityProperty) -> bool {
-        if let Some(global_entity) = *self.global_entity_prop {
-            if let Some(other_global_entity) = *other.global_entity_prop {
-                return global_entity == other_global_entity;
-            }
-            return false;
-        }
-        other.global_entity_prop.is_none()
     }
 
     // Internal
 
-    pub fn set_mutator(&mut self, mutator: &PropertyMutator) {
-        self.global_entity_prop.set_mutator(mutator);
+    pub fn get<E: Copy + Eq + Hash>(
+        &self,
+        converter: &dyn EntityAndGlobalEntityConverter<E>,
+    ) -> Option<E> {
+        match &self.inner {
+            EntityRelation::HostOwned(inner) => inner.get(converter),
+            EntityRelation::RemoteOwned(inner) => inner.get(converter),
+            EntityRelation::RemoteWaiting(_) => {
+                panic!("Not ready to get RemoteWaiting Relation value!");
+            }
+        }
+    }
+
+    pub fn set<E: Copy + Eq + Hash>(
+        &mut self,
+        converter: &dyn EntityAndGlobalEntityConverter<E>,
+        entity: &E,
+    ) {
+        match &mut self.inner {
+            EntityRelation::HostOwned(inner) => {
+                inner.set(converter, entity);
+            }
+            EntityRelation::RemoteOwned(_) | EntityRelation::RemoteWaiting(_) => {
+                panic!("Remote Relations should never be set manually.");
+            }
+        }
+    }
+
+    // Waiting
+
+    pub fn waiting_local_entity(&self) -> Option<LocalEntity> {
+        match &self.inner {
+            EntityRelation::HostOwned(_) | EntityRelation::RemoteOwned(_) => None,
+            EntityRelation::RemoteWaiting(inner) => Some(inner.local_entity),
+        }
+    }
+
+    pub fn waiting_complete(&mut self, converter: &dyn LocalEntityAndGlobalEntityConverter) {
+        match &mut self.inner {
+            EntityRelation::HostOwned(_) | EntityRelation::RemoteOwned(_) => {
+                panic!("Can't complete a RemoteOwned or HostOwned Relation!");
+            }
+            EntityRelation::RemoteWaiting(inner) => {
+                if let Ok(global_entity) =
+                    converter.local_entity_to_global_entity(&inner.local_entity)
+                {
+                    let mut new_impl = RemoteOwnedRelation::new();
+                    new_impl.global_entity = Some(global_entity);
+
+                    self.inner = EntityRelation::RemoteOwned(new_impl);
+                } else {
+                    panic!("Could not find Global Entity from Local Entity! Should only call `waiting_complete` when it is known the converter will not fail!");
+                }
+            }
+        }
+    }
+}
+
+// HostOwnedRelation
+#[derive(Clone)]
+struct HostOwnedRelation {
+    global_entity: Option<GlobalEntity>,
+}
+
+impl HostOwnedRelation {
+    pub fn new() -> Self {
+        Self {
+            global_entity: None,
+        }
+    }
+
+    pub fn write(
+        &self,
+        writer: &mut dyn BitWrite,
+        converter: &mut dyn LocalEntityAndGlobalEntityConverterMut,
+    ) {
+        if let Some(global_entity) = &self.global_entity {
+            let Ok(local_entity) = converter.get_or_reserve_host_entity(global_entity) else {
+                warn!("Global Entity does not Exist! This should not happen.");
+                false.ser(writer);
+                return;
+            };
+            // Must reverse the LocalEntity because the Host<->Remote
+            // relationship inverts after this data goes over the wire
+            let reversed_local_entity = local_entity.to_reversed();
+
+            true.ser(writer);
+            reversed_local_entity.owned_ser(writer);
+            return;
+        }
+        false.ser(writer);
+    }
+
+    pub fn bit_length(&self, converter: &mut dyn LocalEntityAndGlobalEntityConverterMut) -> u32 {
+        let mut bit_counter = BitCounter::new(0, 0, u32::MAX);
+        self.write(&mut bit_counter, converter);
+        return bit_counter.bits_needed();
     }
 
     pub fn get<E: Copy + Eq + Hash>(
         &self,
         converter: &dyn EntityAndGlobalEntityConverter<E>,
     ) -> Option<E> {
-        if let Some(global_entity) = *self.global_entity_prop {
-            if let Ok(entity) = converter.global_entity_to_entity(&global_entity) {
-                return Some(entity);
+        if let Some(global_entity) = self.global_entity {
+            if let Ok(world_entity) = converter.global_entity_to_entity(&global_entity) {
+                return Some(world_entity);
+            } else {
+                warn!("Could not find World Entity from Global Entity, in order to get the EntityRelation value!");
+                return None;
             }
         }
         return None;
@@ -150,12 +215,54 @@ impl EntityProperty {
     pub fn set<E: Copy + Eq + Hash>(
         &mut self,
         converter: &dyn EntityAndGlobalEntityConverter<E>,
-        entity: &E,
+        world_entity: &E,
     ) {
-        if let Ok(new_global_entity) = converter.entity_to_global_entity(entity) {
-            *self.global_entity_prop = Some(new_global_entity);
+        if let Ok(new_global_entity) = converter.entity_to_global_entity(world_entity) {
+            self.global_entity = Some(new_global_entity);
         } else {
-            panic!("Could not find Entity, in order to set the EntityProperty value!")
+            warn!("Could not find Global Entity from World Entity, in order to set the EntityRelation value!");
+            return;
         }
+    }
+}
+
+// RemoteOwnedRelation
+#[derive(Clone)]
+struct RemoteOwnedRelation {
+    global_entity: Option<GlobalEntity>,
+}
+
+impl RemoteOwnedRelation {
+    fn new() -> Self {
+        Self {
+            global_entity: None,
+        }
+    }
+
+    pub fn get<E: Copy + Eq + Hash>(
+        &self,
+        converter: &dyn EntityAndGlobalEntityConverter<E>,
+    ) -> Option<E> {
+        if let Some(global_entity) = self.global_entity {
+            if let Ok(world_entity) = converter.global_entity_to_entity(&global_entity) {
+                return Some(world_entity);
+            } else {
+                warn!("Could not find World Entity from Global Entity, in order to get the EntityRelation value!");
+                return None;
+            }
+        }
+        return None;
+    }
+}
+
+// RemoteWaitingRelation
+#[derive(Clone)]
+struct RemoteWaitingRelation {
+    local_entity: LocalEntity,
+}
+
+impl RemoteWaitingRelation {
+    fn new(local_entity: LocalEntity) -> Self {
+        Self { local_entity }
     }
 }
